Move the parameterized test to RustIterAdapterTest
